Coloquei o <TelaCadastro /> na função renderizarTela, apareceu uma barra de rolagem horitonzal na tela.
