+++ src/components/AppContainer.js
@@ -4,8 +4,7 @@ import CardServicos from './CardServicos/CardServicos';
 import Home from './Home/Home';
 import Header from './Header/Header';
 import TelaCarrinho from './TelaCarrinho/TelaCarrinho';
-import TelaServicos from './TelaServicos/TelaServicos'
-
+import TelaCadastro from './TelaCadastro/TelaCadastro';
 
 export class AppContainer extends Component {
   state = {
@@ -13,7 +12,6 @@ export class AppContainer extends Component {
   }
 
   trocarTela = (tela) => {
-    console.log("TELA:", tela)
     this.setState({ tela: tela })
   }
 
@@ -28,7 +26,7 @@ export class AppContainer extends Component {
     }
     if (this.state.tela === "cadastro") {
       return (
-        <div>Cadastro</div>
+        <TelaCadastro />
       )
     }
     if (this.state.tela === "servicos") {
@@ -47,13 +45,9 @@ export class AppContainer extends Component {
   render() {
     return (
       <div>
-
-        <Header 
-          trocarTela = {this.trocarTela}
+        <Header
+          trocarTela={this.trocarTela}
         />
-
-
-
         {this.renderizarTela()}
 
         {/* O componente EstilosAplicados é só para testar os estilos do Estilos.js e pode ser comentado  */}
