+++ src/components/Botao/Botao.js
@@ -10,3 +10,4 @@ export default class Botao extends React.Component {
     );
   }
 }
+
